Dev2

+++ client/dicebox_client.py
@@ -24,12 +24,12 @@ ramp_frames = 3
 camera = cv2.VideoCapture(camera_port)
 
 
-#camera.set(cv.CV_CAP_PROP_FRAME_WIDTH, 10000);
-#camera.set(cv.CV_CAP_PROP_FRAME_HEIGHT, 10000);
+camera.set(cv.CV_CAP_PROP_FRAME_WIDTH, 1000);
+camera.set(cv.CV_CAP_PROP_FRAME_HEIGHT, 1000);
 
 # 780x650
-camera.set(cv.CV_CAP_PROP_FRAME_WIDTH, 780);
-camera.set(cv.CV_CAP_PROP_FRAME_HEIGHT, 650);
+#camera.set(cv.CV_CAP_PROP_FRAME_WIDTH, 1);
+#camera.set(cv.CV_CAP_PROP_FRAME_HEIGHT, 1);
 
 #60x50
 #camera.set(cv.CV_CAP_PROP_FRAME_WIDTH, 60);
@@ -37,12 +37,30 @@ camera.set(cv.CV_CAP_PROP_FRAME_HEIGHT, 650);
 
 def get_image():
     # read is the easiest way to get a full image out of a VideoCapture object.
-    retval, im = camera.read()
+    retval, camera_capture = camera.read()
     # Our operations on the frame come here
-    im = cv2.cvtColor(im, cv2.COLOR_BGR2GRAY)
+    im = cv2.cvtColor(camera_capture, cv2.COLOR_BGR2GRAY)
     #resized_im = resize_keep_aspect_ratio(im, 255)
 
-    resized_image = cv2.resize(im, (60, 50))
+    height, width = im.shape[:2]
+
+    extra_height = float(height) % 60
+    height_offset = int(extra_height / 2)
+
+    extra_width = float(width) % 50
+    width_offset = int(extra_width / 2)
+
+    # now crop
+    x1 = int(0 + width_offset)
+    y1 = int(0 + height_offset)
+    x2 = int(width - width_offset)
+    y2 = int(height - height_offset)
+    #print("(%i, %i), (%i, %i)" % (y1, y2, x1, x2))
+    cropped_image = im[y1:y2, x1:x2]
+
+
+    resized_image = cv2.resize(cropped_image, (60, 50))
+    # resized_image = cv2.resize(im, (60, 50))
     height, width = resized_image.shape[:2]
 
     height = float(height)
@@ -50,7 +68,7 @@ def get_image():
     #print("height: (%f)" % height)
     #print("width: (%f)" % width)
 
-    return resized_image
+    return camera_capture, resized_image
 
 
 def resize_keep_aspect_ratio(input, desired_size):
@@ -124,24 +142,42 @@ for d in jdata:
     #print("%s:%s" % (d, jdata[d]))
     server_category_map[d] = jdata[d]
 
-print(server_category_map)
+CURRENT_EXPECTED_CATEGORY_INDEX=1
+MAX_EXPECTED_CATEGORY_INDEX = len(server_category_map)
+MISCLASSIFIED_CATEGORY_INDEX = True
+
+KEEP_INPUT = False
+ONLY_KEEP_MISCLASSIFIED_INPUT = True
+
+SERVER_ERROR = False
+
+#print(server_category_map)
 ###############################################################################
 # main loop
 ###############################################################################
 while (True):
     # Take the actual image we want to keep
     #camera_capture, resized_image  = get_image()
-    camera_capture  = get_image()
+    camera_capture, resized_image = get_image()
     filename = datetime.now().strftime('capture_%Y-%m-%d_%H_%M_%S_%f.png')
 
     # A nice feature of the imwrite method is that it will automatically choose the
     # correct format based on the file extension you provide. Convenient!
-    cv2.imwrite('./tmp/%s' % filename, camera_capture)
+    cv2.imwrite('./tmp/%s' % filename, resized_image)
 
     with open('./tmp/%s' % filename, 'rb') as file:
         file_content = file.read()
-    os.remove('./tmp/%s' % filename)
-    #os.rename('./tmp/%s' % filename, './data/1d4/%s' % filename)
+
+    if KEEP_INPUT and not SERVER_ERROR:
+        #print(MISCLASSIFIED_CATEGORY_INDEX)
+        #print(ONLY_KEEP_MISCLASSIFIED_INPUT)
+        if not MISCLASSIFIED_CATEGORY_INDEX and ONLY_KEEP_MISCLASSIFIED_INPUT:
+            os.remove('./tmp/%s' % filename)
+        else:
+            os.rename('./tmp/%s' % filename, './tmp/%s/%s' % (server_category_map[str(CURRENT_EXPECTED_CATEGORY_INDEX-1)], filename))
+    else:
+        os.remove('./tmp/%s' % filename)
+
     base64_encoded_content = file_content.encode('base64')
 
     outjson = {}
@@ -150,7 +186,7 @@ while (True):
     json_data = json.dumps(outjson)
 
     prediction = {}
-
+    category = {}
     #print ('sending over the wire: %s' % json_data)
     headers = {
         'Content-type': 'application/json',
@@ -164,23 +200,37 @@ while (True):
         response = requests.post('http://127.0.0.1:5000/api/prediction', data=json_data, headers=headers)
         if response is not None:
             if response.status_code != 500:
+                SERVER_ERROR = False
                 if 'prediction' in response.json():
                     prediction = response.json()['prediction']
                     category = server_category_map[str(prediction)]
                     #print("%s" % prediction)
                     #print("%s" % category)
     except:
-        print('.')
+        #print('.')
+        SERVER_ERROR = True
         #raise
 
     #print prediction
 
-    #for key, value in sorted(prediction.iteritems(), key=lambda (k, v): (v, k)):
-    #    #print "%s: %s" % (key, value)
-    #    cv2.putText(camera_capture, "%s: %s" % (key, value), (5, 20 + 20 * i), font, 0.5, (255, 255, 255), 1)
-    #    i -= 1
-    #    if i < 0:
-    #        i = 4
+    # training_category = '1d4_4'
+    # if category == training_category:
+    #     os.rename("./tmp/%s" % filename, "./tmp/%s/%s" % (category, filename))
+    # else:
+    #     print('misclassified')
+    #     os.rename("./tmp/%s" % filename, "./tmp/misclassified/%s" % filename)
+
+
+    #print(MAX_EXPECTED_CATEGORY_INDEX)
+    #print(CURRENT_EXPECTED_CATEGORY_INDEX)
+    #print(server_category_map[str(CURRENT_EXPECTED_CATEGORY_INDEX-1)])
+
+    if category == server_category_map[str(CURRENT_EXPECTED_CATEGORY_INDEX-1)]:
+        MISCLASSIFIED_CATEGORY_INDEX = False
+        #print(False)
+    else:
+        MISCLASSIFIED_CATEGORY_INDEX = True
+        #print(True)
 
     cv2.namedWindow('dice box', cv2.WINDOW_NORMAL)
     # lets make a pretty output window
@@ -189,12 +239,40 @@ while (True):
     #cv2.imshow('dice box', camera_capture)
 
     #cv2.putText(camera_capture, "%s" % category, (5, 15), font, 0.5, (255, 255, 255), 1)
-    cv2.putText(camera_capture, "%s" % category, (5, 15), font, 0.3, (255, 255, 255), 1)
-    cv2.imshow('dice box', camera_capture)
 
-    if cv2.waitKey(1) & 0xFF == ord('q'):
+    # im[y1:y2, x1:x2]
+    output_display = camera_capture
+    height, width = output_display.shape[:2]
+    output_display[height-50:height, 0:60] = cv2.cvtColor(resized_image, cv.CV_GRAY2RGB)
+
+    output_label = "[pred %s/exp %s][match? %r][record? %r][only keep misclassified? %r][server error? %r]" % (category, server_category_map[str(CURRENT_EXPECTED_CATEGORY_INDEX-1)], not MISCLASSIFIED_CATEGORY_INDEX, KEEP_INPUT, ONLY_KEEP_MISCLASSIFIED_INPUT, SERVER_ERROR)
+
+    cv2.putText(output_display, output_label, (5, 15), font, 0.3, (255, 255, 255), 1)
+
+    cv2.imshow('dice box', output_display)
+
+    input_key = cv2.waitKey(1)
+
+    if input_key & 0xFF == ord('q'):
         break
 
+    if input_key & 0xFF == ord('c'):
+        if CURRENT_EXPECTED_CATEGORY_INDEX >= MAX_EXPECTED_CATEGORY_INDEX:
+            CURRENT_EXPECTED_CATEGORY_INDEX = 1
+        else:
+            CURRENT_EXPECTED_CATEGORY_INDEX += 1
+
+    if input_key & 0xFF == ord('z'):
+        if KEEP_INPUT == True:
+            KEEP_INPUT = False
+        else:
+            KEEP_INPUT = True
+
+    if input_key & 0xFF == ord('b'):
+            if ONLY_KEEP_MISCLASSIFIED_INPUT == True:
+                ONLY_KEEP_MISCLASSIFIED_INPUT = False
+            else:
+                ONLY_KEEP_MISCLASSIFIED_INPUT = True
 
 ###############################################################################
 # cleanup
